Removing docker-compose and improvements.

+++ integration-testing/docker_run_tests.sh
@@ -15,12 +15,17 @@ MAX_NODE_COUNT=10
 
 cleanup() {
     # Eliminate this for next run
-    rm docker-compose.yml
+    if [[ -f "docker-compose.yml" ]]; then
+        rm docker-compose.yml
+    fi
 
     for num in $(seq 0 $MAX_NODE_COUNT)
     do
-        docker network rm -f cl-${TAG_NAME}-${num}
+        # Network might get tore down with docker-compose rm, so "|| true" to ignore failure
+        docker network rm cl-${TAG_NAME}-${num} || true
     done
+
+    docker network prune --force
 }
 trap cleanup 0
 
@@ -33,4 +38,4 @@ done
 # Using ||TAG|| as replacable element in docker-compose.yml.template
 sed 's/||TAG||/'"${TAG_NAME}"'/g' docker-compose.yml.template > docker-compose.yml
 
-docker-compose up && docker-compose rm -f
+docker-compose up && docker-compose rm --force --stop
